+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-28 15:33:40
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-08-29 15:17:06
+// Commit: docs(core): update readme and API contracts for core
 
 class coreComponentState {
-  final String commitRef = '3992';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-08-28T15:33:40+05:00');
+  final String commitRef = '4031';
+  final String lastMessage = 'docs(core): update readme and API contracts for core';
+  final DateTime timestamp = DateTime.parse('2026-08-29T15:17:06+05:00');
 }